Split test_reporter.py (1280 LOC) into four focused modules
tests/bench/test_reporter.py exceeded the 1000 LOC block threshold for
the local model context window. Split into:
  - test_reporter_eligibility.py  (291 LOC)  _is_eligible, _percentile, _cv
  - test_reporter_ranking.py      (361 LOC)  print_ranking integration
  - test_reporter_apc.py          (337 LOC)  APC speedup, concurrency efficiency
  - test_reporter_compare.py      (361 LOC)  _metric_delta, print_compare_table

All 134 tests pass after split.

Closes pre-epic-close split gate for WOR-209.
